refactor: separate views into folders
